bgta updates; use common bubble creator; fix bddt bubble positions.
Former-commit-id: 7020a424122a9485eb2045138d70c1141ca83cde

+++ javasrc/edu/brown/cs/bubbles/bump/BumpRunManager.java
@@ -1514,14 +1514,17 @@ private class StackFrame implements BumpStackFrame {
 		     for_file = File.createTempFile("BUBBLES_" + xnm,".java");
 		     source_map.put(fnm,for_file);
 		     byte [] data = IvyXml.stringToByteArray(IvyXml.getTextElement(xml,"SOURCE"));
-		     FileOutputStream fos = new FileOutputStream(for_file);
-		     fos.write(data);
-		     fos.close();
+		     if (data == null) for_file = null;
+		     else {
+			FileOutputStream fos = new FileOutputStream(for_file);
+			fos.write(data);
+			fos.close();
+		      } 
 		   }
 		  catch (IOException e) {
 		     BoardLog.logE("BUMP","Problem writing source file: " + e,e);
 		   }
-		  for_file.deleteOnExit();
+		  if (for_file != null) for_file.deleteOnExit();
 		}
 	     }
 	  }
